[ORC-218] Cache timezone information in the library.
Right now, for every lookup of a timezone, the library will go to
disk and parse the timezone file. While the results are cached,
doing these system calls should be avoided in environments with
restricted system calls.

This patch loads all the time zone files into the binary at
build time and instead of accessing disk for the TZ
information will simply load the buffer from memory. In addition,
this allows the user to compile the library with a static
timezone to be assumed for local use. This comes in handy to avoid
yet another set of system calls to assume the local timezone.

The following CMake flags control this feature:

  * EMBEDED_TZ_DB - disables this feature if set to OFF
  * STATIC_TZ=Val - sets a specific timezone