ORC-600: Fix StringDictionaryColumnReader to update index buffer correctly
When we are scanning with StringDictionaryColumnReader an encoded batch,
ORC crashes if we want to read more elements than the current capacity
of the batch. It can occur eg. with string lists where it's possible
that the offsets defined in the list overflows in the index buffer of
the EncodedStringDictionaryBatch.

This commit provides a fix by defining a resize method for
EncodedStringDictionaryBatch that also resizes the index buffer.

Tests:
 - Added a test that would cause ORC to crash without the fix.