Int128: Undefined behaviour in left-shift
Left-shifting into the sign bit is Undefined Behaviour and UBSAN reports it executing ORC tests. If shiftint into the sign it is desied, the workaround is doing the shift in unsigned types, and casting it to signed. Casting an unsigned type to signed when the value is not representable by the signed type is Implementation Defined, but nearly all implementations just maintain all bits.

With this patch, all tests from "orc-test" pass correctly with no reports from UBSAN.