ORC-648: Add GitHub Action for Java8/11 test coverage
This PR aims to add `GitHub Action` for explicit Java 8/11 test coverage. Since this is an independent CI setup, this will not interfere with the existing CIs.

Apache ORC is currently using
- Travis CI for testing various `clang` versions on Trusty and Mac with JDK7.
- Appveyor CI for testing Visual Studio on Windows OS.

This is tested here.
- https://github.com/dongjoon-hyun/orc/pull/3

![Screen Shot 2020-07-16 at 9 56 22 PM](https://user-images.githubusercontent.com/9700541/87750050-3aec7280-c7af-11ea-9e2e-adbdf5bb7a9c.png)

After this PR, it seems that we need to file Apache INFRA JIRA issue to enable `GitHub Action` on ORC project.