ORC-639: Improve zstd performance (#537)
### What changes were proposed in this pull request?

This PR aims to improve ZSTD performance in branch-1.6. This is a backport of #511

### Why are the changes needed?

Zstd's "Simple API" is used which initializes the compression/decompression context each time. This includes time to initialize all compression/decompression tables. Using the "Explicit context" API improves the performance significantly as compression/decompression tables are constructed once and reused each time "decompress" or "doBlockCompression" is called. We've noticed more than 15% time improvement in some applications with this change, so it seems that any ORC user using Zstd can benefit from this change.

### How was this patch tested?

Pass the UT