ORC-689: Add GitHubAction job to publish snapshot (#573)
### What changes were proposed in this pull request?

This PR aims to add `GitHub Action` job to publish snapshot like the following. The following is generated by this patch from my fork `master` branch.

- https://repository.apache.org/content/repositories/snapshots/org/apache/orc/orc-core/1.7.0-SNAPSHOT/

### Why are the changes needed?

Currently, Apache ORC community doesn't provide snapshot artifacts. It would be great if we provide it in order to help the downstream projects to test it in advance.

### How was this patch tested?

This is tested in my ORC fork and we need INFRA team's credential setup.

- https://github.com/dongjoon-hyun/orc/runs/1468464962 (I merged to my `master` branch and removed it after testing)
- https://issues.apache.org/jira/browse/INFRA-21150 (Upload snapshots from GitHub Actions)