ORC-714: Remove MRUnit and its usage (#612)
### What changes were proposed in this pull request?

This PR aims to remove `mrunit` and its dependency and usage. This will unblock us toward new Java versions like Java 17.
```
[INFO] +- org.apache.mrunit:mrunit:jar:hadoop2:1.1.0:test
[INFO] |  +- org.powermock:powermock-core:jar:1.5.1:test
[INFO] |  |  +- org.powermock:powermock-reflect:jar:1.5.1:test
[INFO] |  |  \- org.javassist:javassist:jar:3.18.0-GA:test
[INFO] |  +- org.powermock:powermock-api-mockito:jar:1.5.1:test
[INFO] |  |  \- org.powermock:powermock-api-support:jar:1.5.1:test
[INFO] |  \- org.powermock:powermock-module-junit4:jar:1.5.1:test
[INFO] |     \- org.powermock:powermock-module-junit4-common:jar:1.5.1:test
```
### Why are the changes needed?

We are using the last version, `mrunit`, which is released at `2014/06/06` and is retired `2016/04/30`. We had better remove this and use Hadoop's minicluster or alternatives.

### How was this patch tested?

Pass the CIs because this is a removal of test coverage. ORC-715 will add a test coverage back.