ORC-766: Generalize the docker scripts to handle build arguments. (#658)
### What changes were proposed in this pull request?

This replaces the build matrix loops in the docker reinit.sh and run-all.sh scripts with expanding the
versions that we want to build in os-list.txt. The pattern for the build args is based on the name:

ubuntu20 builds ubuntu 20 with default parameters jdk=8 and cc=gcc
ubuntu20_jdk=11_cc=clang builds ubuntu 20 with jdk=11 and cc=clang

Those names are carried through to the docker tags with s/=/-/.

### How was this patch tested?

I ran reinit.sh and run-all.sh and both ran fine. The docker images default to the proper c++ and java compilers.