MINOR: Fix RLEv1 encoding examples in ORCv0/v1/v2 docs (#676)
Change the literals-encoding example to [2,3,6,7,11]. Either [2,3,4,7,11] or first 5 prime numbers [2,3,5,7,11] has run, and can not be used as a literals-encoding example.

------

The first 5 prime numbers are [2,3,5,7,11], and there is a run. The encoded data is
- 2 [0xff, 0x02]
- 3,5,7 (delta=2,length=3,base=3) [0x00, 0x02, 0x03]
- 11 [0xff, 0xb]

And if we use example [2,3,4,7,11] as the spec written, there is also a run .The encoded data is
- 2,3,4 (delta=1, length=3, base=2) [0x00, 0x01, 0x02]
- 7,11  [0xfe, 0x7, 0xb]

------
And I have verified the encoded data with the following C++ code. It used RLEEncoderv1 to encode [2,3,5,7,11] and dump data to a local file called `rle.data`.

```c++

int main() {
    auto pool = getDefaultPool();
    auto outputStream = writeLocalFile("rle.data");
    auto buf = new BufferedOutputStream(*pool, outputStream.get(), 1024, 32);
    std::unique_ptr<BufferedOutputStream> pbuf(buf);
    auto encoder = new RleEncoderV1(std::move(pbuf), false);
    encoder->write(2);
    encoder->write(3);
    encoder->write(5);
    encoder->write(7);
    encoder->write(11);
    encoder->flush();
    return 0;
}
```

And then I use `od -A n -t x1 rle.data` to see data in hex format. The result is following

```
 ff 02 00 02 03 ff 0b
```

(cherry picked from commit 0336d24a980b8f10d49b1482c505f9c899a1eae7)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>