ORC-777: Make the vectorized row batch size configurable in MR record readers and writers (#674)
### What changes were proposed in this pull request?

The changes in this PR create overloaded constructors in MapreduceRecordReader/Writer classes that accept the rowbatch size as an argument. In order to use the overloaded constructor, the caller is expected to override the getRecordWriter() method (or getRecordReader()), as exemplified in the TestMapreduceOrcOutputFormat class.

### Why are the changes needed?

The change is needed to provide greater configurability of the ORC MR readers/writers.
Current behavior of OrcInput(Output)Format is to create an ORC reader/writer with the default row batch size of 1024. This may need to be overridden in certain scenarios to prevent OOMs during reads/writes.

### How was this patch tested?

Added unit tests in TestOrcOutputFormat and TestMapreduceOrcOutputFormat, where we extend the OrcOutputFormat class to override the getRecordWriter() method. The change to override the row batch size config was also tested by running a MR job that used the snapshot version of the orc-mapreduce library.

(cherry picked from commit fc4cf4c57fa30c3eb684cd63d7658d966821852c)