ORC-797: Allow writers to get the stripe information. (#698)
### What changes were proposed in this pull request?

It adds a new method to the Writer that lets the user get information about the stripes of the file.

### Why are the changes needed?

Currently Iceberg has to reopen the file after writing the file and we'd like to short circuit that.

### How was this patch tested?

Added a unit test.

(cherry picked from commit 152304ecd5175c22ce42fe72e9d9199d0adb75b8)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>