ORC-799: Remove Ubuntu 16 docker test (#699)
### What changes were proposed in this pull request?

This PR aims to delete Ubuntu 16 docker test.

### Why are the changes needed?

According to [Ubuntu LTS lifecycle](https://wiki.ubuntu.com/Releases), `End of Standard Support` of `Ubuntu 16` was April 2021. Although `End Of Life` is still 2024, we can delete it from our test environment for Apache ORC 1.7.0.

### How was this patch tested?

N/A (This is a removal of test coverage).