ORC-826: Do Not Use Collection Contains/Get
1. What changes were proposed in this pull request?

Discovered multiple places in the code where a Collection is accessed with a 'contains' and a 'get'. However, this is superfluous as a 'get' will return a null value if the Collection doesn't contain the item. This can be utilized to save multiple trips into the same map.

2. Why are the changes needed?

Small performance, less code

3. How was this patch tested?

No change in behavior so existing unit test suffice.

Closes #731

Signed-off-by: William Hyun <william@apache.org>