ORC-831: Do Not Copy String When Flushing Dictionary (#736)
### What changes were proposed in this pull request?

The code to flush first loads (copies) the data into a Text object and then writes it to the direct stream. Instead, pass the directStream to the dictionary and have it write directly to the stream instead of an intermediate Text object.

### Why are the changes needed?

Performance.

### How was this patch tested?

No change in behavior so using existing unit tests.