ORC-830: Do Not Copy String When Adding to StringHashTableDictionary (#735)
### What changes were proposed in this pull request?
When there is a collision adding a value into a StringHashTableDictionary, a temp Text object is created and then each value in the byte array is copied into the temp Text until a match is found (or worst-case scenario, a match is not found and every value is loaded).

Instead of loading (copying) the values, just compare directly against the byte array without copying the data into a intermediate (temp) buffer.

### Why are the changes needed?
Performance.  StringTreeWriter#writeBatch() consumes some number the cycles, much of which is spent in getText().  By Changing this getText() implementation (removing the copy), the number of cycles consumed is decreased measurably.

### How was this patch tested?
No functionality change, utilized existing unit tests.