ORC-959: [C++] Fix crash of SargsApplier in finding nested field names (#877)
### What changes were proposed in this pull request?

This patch fixes a crash in the C++ reader when SearchArgument is created on nested columns of collection types (e.g. array/map). Only STRUCT type has field names for subtypes. So we should check the current node type before checking any field names. JIRA description contains details on how to reproduce the crash.

### Why are the changes needed?

Same as the above.

### How was this patch tested?

Added unit tests in c++/test/TestSargsApplier.cc to reproduce the crash and verify the fix on ARRAY and MAP types.