ORC-980: Filter processing respects the case-sensitivity flag (#893)
The filter processing incorrectly ignores the case-sensitivity flag for schema supplied by the reader. This fixes this bug and uses the flag in determining field and vector information.

In the absence of this the read incorrectly fails by performing a case-sensitive match even when case-insensitive match was requested by the reader.

Unit Tests were added to verify the failure and subsequently the fix.

(cherry picked from commit 59c078d89d0fe2d2921b37113731cc032efc472e)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>