ORC-980: Filter processing respects the case-sensitivity flag (#893)
### What changes were proposed in this pull request?

The filter processing incorrectly ignores the case-sensitivity flag for schema supplied by the reader. This fixes this bug and uses the flag in determining field and vector information.

### Why are the changes needed?

In the absence of this the read incorrectly fails by performing a case-sensitive match even when case-insensitive match was requested by the reader.

### How was this patch tested?

Unit Tests were added to verify the failure and subsequently the fix.