ORC-984: Save the software version that wrote each ORC file. (#904)
### What changes were proposed in this pull request?

I add a string to the file footer that records the version that wrote the file. We already had recorded the implementation that wrote the file as Footer.writer. I added a method that combines these two fields in the reader to produce a user facing string that describes the software version.

I also add a field to the meta data tool to show the version that wrote the file.

Because of that change and the fact that the files change size based on whether the ORC version is a snapshot or not, I had to extend the tests for TestFileDump to allow some slop for the size and to ignore the file version.

### How was this patch tested?

It passes the unit tests after I updated the tools tests.