ORC-1023: Support writing bloom filters in ConvertTool (#933)
### What changes were proposed in this pull request?

This PR adds an option to the java tool ConvertTool to specify which columns it should generate bloom filters.

### Why are the changes needed?

While debugging an issue, I need to generate an ORC file with bloom filters using the Java APIs. The ConvertTool is easy to use but it doesn't generate bloom filters. It'd be helpful to add an option for it.

### How was this patch tested?

Didn't find any existing tests on ConvertTool. So I manually tested it and verified the bloom filters are generated.

(cherry picked from commit 7c451371add581472c05164c0c0951bb64963116)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>