Fix test failure in RELEASE builds when using GCC 9
Previously the inline method getLongHash is defined in BloomFilter.cc.
To test it we include BloomFilter.cc in TestBloomFilter.cc. GCC 9 seems
to have a bug in optimizing the inline method that it incorrectly
removes some instructions. The failures are seen using GCC 9.3 or 9.4.

This commit move the getLongHash method to the header file. So we don't
need to include the source file, which work around the issue.