ORC-1029: ServiceLoader is not thread-safe, avoid concurrent calls (#940)
### What changes were proposed in this pull request?

The previous code reuses LOADER to build DataMask, but the ServiceLoader class is not thread-safe and will cause an exception if called concurrently.
Checking the source code, I found that ServiceLoader.load is a lazy method, which is relatively lightweight and only records the relevant information until the iterator starts reading before loading the class, so I think it is acceptable not to reuse it.

### Why are the changes needed?

Avoid concurrent calls leading to exceptions.

### How was this patch tested?

N/A

(cherry picked from commit 19f184acbd02e2ec06f137d24ebefdead709da0c)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>