ORC-1034: Fix the indexOf algorithm in `FileDump` (#943)
This matching algorithm is wrong when i does not backtrack after a failed match in the middle. As a simple example data = OOORC, pattern= ORC, index = 1, this algorithm will return -1.

This pr aims to fix the indexOf algorithm.

indexOf is used to find the ORC file ending identifier to recover the file, it is important to ensure that the method is correct.

Add UT.

(cherry picked from commit 65ce439fc7cce84f12657e73659165d20a306bd9)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>