ORC-1081: Fix heap-use-after-free in SearchArgumentBuilderImpl::end() (#998)
### What changes were proposed in this pull request?

This PR fixes the heap-use-after-free issue in `SearchArgumentBuilderImpl::end()`. `std::deque::pop_front` will delete the element. We should call it after using the element.

### Why are the changes needed?

This crashes the test when building ORC with AddressSanitizer. The ASAN report is attached in the JIRA.

### How was this patch tested?

After this fix, I'm able to run all existing tests with AddressSanitizer enabled.