ORC-1082: Improve FileDump and JsonFileDump to be robust on missing column statistics (#1003)
### What changes were proposed in this pull request?

This pr is aims to fixing code in the FileDump and JsonFileDump commands that determines whether a column statistic exists or not.

### Why are the changes needed?

Before we can get the ColumnStatistics from RowIndex we need to determine whether it exists or not.

`entry.getStatistics()` does not return null at any time, and will return a default object when it is not set.

The current impl is wrong
```java
      OrcProto.ColumnStatistics colStats = entry.getStatistics();
      if (colStats == null) {
        buf.append("no stats at ");
      }
```

### How was this patch tested?

This can be tested with the following code

```java
  @Test
  public void test() {
    OrcProto.RowIndexEntry entry = OrcProto.RowIndexEntry.newBuilder().addPositions(0).build();
    assertFalse(entry.hasStatistics());
    assertNotNull(entry.getStatistics());
  }
```
But since this is just testing Protobuf generated code, I'm not sure if I should add it to the unit tests.

On the other hand, our proto is defined as follows, statistics is optional

https://github.com/apache/orc/blob/7d2838b4b38111bc1bfa00592d84a8b36286b9f2/proto/orc_proto.proto#L102-L105

Our orc writer implementation, however, will definitely write statistics, and there is no optional configuration, so it cannot be verified by the files written by orc

https://github.com/apache/orc/blob/7d2838b4b38111bc1bfa00592d84a8b36286b9f2/java/core/src/java/org/apache/orc/impl/writer/TreeWriterBase.java#L322-L330
(cherry picked from commit 043a827783c2d21ebf7f28355f66f0bb97db0834)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>