ORC-1004: Java ORC writer supports the selection vector (#994)
### What changes were proposed in this pull request?

This PR is to enable the Java ORC writer to respect the selection vector from VectorizedRowBatch. The implementation tries to find each longest batch that is continuously selected to get the best performance.

### Why are the changes needed?

Currently the ORC writer doesn't support the selected vector. This could cause clients that expect it to be supported to get trash rows in the output.

### How was this patch tested?

Enabled UT TestSelectedVector.java