ORC-1095: Deprecate `UnknownFormatException` (#1016)
### What changes were proposed in this pull request?

This PR aims to deprecate `UnknownFormatException` at Apache ORC 1.7.3.

### Why are the changes needed?

This extends `IOException` to expose more data and used once here. However, the downstream projects are not using this information because `checkOrcVersion` returns `IOException`.  The error message is enough.

https://github.com/apache/orc/blob/80b98126cf8a15d9d44c67a9aabfabfdcc7f913e/java/core/src/java/org/apache/orc/impl/ReaderImpl.java#L517-L525

This PR replaces it with `IOException` with the same message generated by `UnknownFormatException` previously.

https://github.com/apache/orc/blob/80b98126cf8a15d9d44c67a9aabfabfdcc7f913e/java/core/src/java/org/apache/orc/UnknownFormatException.java#L33-L35

### How was this patch tested?

Pass the CIs.