ORC-1078: Row group end offset doesn't accommodate all the blocks (#996)
### What changes were proposed in this pull request?
To calculate the row group end offset based on the buffer size. Current implementation assumes stretching the slop by a constant factor of 2 to accommodate the compression block but thats not always the case when smaller buffer sizes are used.

### Why are the changes needed?
When compression size is smaller than 2048 and two blocks are not enough for all the values left in last row group, we might end up seeking outside of range.

### How was this patch tested?
Unit test added TestRecordReaderImpl#testSmallCompressionSizeOrc

Change-Id: Iff6b0468f582f09b1da7120d6154daf6aa7201f1