ORC-1098: [C++] Support specifying type ids or column names in cpp tools (#1020)
### What changes were proposed in this pull request?

This is a follow-up task of #921. Currently we have options for the tools to work on specified top-level column fields. However, ACID ORC files usually have nested structure. We need the type ids to specify nested columns. As an extension, adding support for column names will also be helpful. So we don't need to manually convert column names to type ids. Also reports the valid values when an invalid column name is given.

This PR extracts the option parsing codes into ToolsHelper. So similiar cpp tools can share the same option set.

### Why are the changes needed?

It makes the tools more useful in practice.

### How was this patch tested?

Added unit tests for the new options.