ORC-1099: [C++] Extend ReadIntent to support MAP and UNION type (#1024)
### What changes were proposed in this pull request?
This patch is a follow-up of ORC-450. Similar to LIST, we can also avoid
the materialization of MAP and UNION column elements by specifying
ReadIntent_OFFSETS in selection.

This patch also adds a requirement that subtypes of UNION should be all
selected or not at all. UnionColumnReader will crash if only part of the
subtypes were selected. This patch modifies ColumnSelector to silently
select all the UNION subtypes if they were partially selected.


### Why are the changes needed?
This patch helps avoid materialization of MAP and UNION elements,
such as in the case of "select count(*)" query.


### How was this patch tested?
- Added several unit tests under TestReadIntent.*