ORC-1127: [C++] add missing version of UNSTABLE-PRE-2.0 (#1064)
### What changes were proposed in this pull request?

This adds the UNSTABLE-PRE-2.0 version in the c++ client. It's part of #1062 but now extracted for easy backporting.
Ref: https://github.com/apache/orc/pull/1062#discussion_r824362393

### Why are the changes needed?

Without this change, the C++ client will read the UNSTABLE-PRE-2.0 version as 1.9999, which is inconsistent with the Java client. See examples in the JIRA description.

### How was this patch tested?

Add unit test in TestFileMetadata.cc with a test file generated by Hive.