ORC-1198: Add a new `PhysicalFsWriter` constructor with `FSDataOutputStream` parameter
### What changes were proposed in this pull request?
Add stream parameter constructor for PhysicalFsWriter.

### Why are the changes needed?
PhysicalFsWriter implementation works on the basis of a Path, but Flink's bulk writer based on stream.
In order to integrate with flink more elegantly,  I think a constructor with stream parameter should be added to PhysicalFsWriter

### How was this patch tested?
It won't change behavior of PhysicalFsWriter and passed all existed test cases.

Closes #1153 from liujiawinds/main.

Authored-by: Jia Liu <liujiawinds@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>