ORC-1204. ORC MapReduce writer to flush when long arrays
This adds a new configuration for the row by row writer that sets a limit of how many items will be buffered in the VectorizedRowBatch before it is sent to the ORC Writer. This does not change the behavior of the core writer, where the application already has control over the size of the batch.

We are getting OOM when writing rows with long arrays with the row-by-row writer. Heap dumps show that it is the ColumnVector inside the array that is taking large amounts of memory.

Updated a couple unit tests and added two more.

Closes #1161 from omalley/orc-1204.

Authored-by: Owen O'Malley <oomalley@linkedin.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>