ORC-961: [C++] expose related metrics of the reader
### What changes were proposed in this pull request?
This patch keeps track of the time spent and the number of calls to each module. The current metrics mainly include the decompression time and number of calls, decoding time and number of calls and total elapsed time.

### Why are the changes needed?
It exposes the relevant metrics of the reader, so that the user can visually see the metrics of each module which includes decompression, decoding.

### How was this patch tested?
The orc-scan tool with the -m parameter can output relevant metrics.

Closes #1158 from coderex2522/ORC-961.

Authored-by: coderex2522 <rex6817@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>