ORC-1288: [C++] fix the bug that invalid memory freeing with zlib compression
### What changes were proposed in this pull request?
This bug can be triggered when writing a orc file that contains multiple stripes, and each stripe contains columns with no null values. The bug is caused by the fact that the member variables in class CompressionStream is not reset after the suppress function is called.

### Why are the changes needed?
The patch can fix the bug about invalid memory freeing with multiple compression kinds.

### How was this patch tested?
Added new UT WriterTest.suppressPresentStreamWithCompressionKinds.

Closes #1282 from coderex2522/ORC-1288.

Authored-by: coderex2522 <rex6817@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>