ORC-1309: Eliminate the redundant `{}` in the Case syntax
### What changes were proposed in this pull request?
This pr aims to eliminate the redundant `{}` in the case syntax.

### Why are the changes needed?
The code has been modified in ORC-1306 to pass the check. But the indentation in the case syntax is weird. The patch for the 1.8 branch merge already fixes it. So you need to fix the main branch separately pr.

### How was this patch tested?
Checking with maven-check-plugin.

Closes #1312 from guiyanakuang/ORC-1309.

Authored-by: Guiyanakuang <guiyanakuang@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>