ORC-1375: Cancel old running ci tasks when a pr has a new commit
### What changes were proposed in this pull request?

This pr is aimed to cancel old running ci tasks when a pr has a new commit.

### Why are the changes needed?

The old tasks were based on outdated code which wasted resources and was pointless. Canceling tasks makes for a more efficient use of resources.

### How was this patch tested?

It doesn't seem to work immediately before merging, I verified this in other project.

Closes #1415 from guiyanakuang/cancel_obsolete_test.

Authored-by: Yiqun Zhang <guiyanakuang@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>