ORC-1362: [Java] Enable ENABLE_INDEXES (#1387)
### What changes were proposed in this pull request?
Enable the functionality behind `orc.create.index`.

### Why are the changes needed?
In https://github.com/apache/orc/pull/1373, we disabled the configuration "orc.create.index" because the orc reader cannot work well when orc.create.index=false.

During the write of a file with `orc.create.index=false` and `orc.row.index.stride > 0` no indexes are created in the ORC file. The `orc.row.index.stride` is persisted into the file metadata. During a read that requires the use of indexes e.g. in the presence of search arguments we get a read failure as the reader was expecting indexes based on the value of `orc.row.index.stride` but none exist.

Two changes are proposed here:
* Enable consistent writes i.e. assume orc.row.index.stride=0 if orc.create.index=false or if orc.row.index.stride >0 then assume orc.create.index=true
* Enable the reader to handle the absence of row index information. During the process of selecting row groups in a stripe if the index information is seen as missing then we assume that all row groups should be read as the index information is missing.

### How was this patch tested?
UT