ORC-1384 Fix ArrayIndexOutOfBoundsException when reading dictionary stream bigger then dictionary
### What changes were proposed in this pull request?
Avoid  ArrayIndexOutOfBoundsException when reading dictionary stream bigger then dictionary. Check the size of the dictionary and input and read only the min of those.

### Why are the changes needed?
In Hive when reading with LLAP data is read in 4kB blocks which leads to ArrayIndexOutOfBoundsException when the dictionary is smaller.

### How was this patch tested?
It is tested with HIVE's qtest, since here we do not have the necessary subclasses.

This closes #1384