ORC-1484: Use `JIRA_ACCESS_TOKEN` in `merge_orc_pr.py`
### What changes were proposed in this pull request?

This PR aims to use `JIRA_ACCESS_TOKEN` instead of a pair of JIRA_USERNAME and JIRA_PASSWORD.

### Why are the changes needed?

To reduce the security vulnerability for the committers, we had better use `Access Tokens` instead of user ID and password.

Personal Access Tokens are available here.
- https://issues.apache.org/jira/secure/ViewProfile.jspa

### How was this patch tested?

Manual.

I used this script when I merge #1586 .

Closes #1589 from dongjoon-hyun/ORC-1484.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>