ORC-1487: Enable `checkstyle` on `src/test` with `checkstyle-suppressions.xml`
### What changes were proposed in this pull request?

This PR aims to enable `checkstyle` on `src/test` directories to validate the recent PRs like #1590 .

### Why are the changes needed?

To help the recent community work, we permanently declare the following rules as exception in test source code.

```xml
     <suppress checks="LineLength" files="src/test/*"/>
     <suppress checks="NewlineAtEndOfFile" files="src/test/*"/>

     <suppress checks="UnusedImports" files="src/test/*"/>
     <suppress checks="AvoidStarImport" files="src/test/*"/>
     <suppress checks="CustomImportOrder" files="src/test/*"/>
```

The other suppressed rules will be removed after we finish the test code clean-ups.

### How was this patch tested?

Pass the CIs.

Closes #1591 from dongjoon-hyun/ORC-1487.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>